fix: hardens job execution and space copy
Improves container and cluster runtime reliability with more portable decoding, safer memory detection, per-run log capture, and more defensive output handling.

Prevents silent failures when loading jobs or copying spaces by surfacing database errors, validating missing source data, and correcting the returned seed type. Adds a lightweight status lookup to support safer status transitions.

+++ starexec-app/src/main/resources/db/migration/R__procedures_and_views.sql
@@ -7874,7 +7874,7 @@ $$ LANGUAGE plpgsql;
 -- Author: Tyler Jensen
 DROP FUNCTION IF EXISTS starexec.GetSpaceJobsById CASCADE;
 CREATE OR REPLACE FUNCTION starexec.GetSpaceJobsById(_spaceId INT)
-RETURNS TABLE(id INT, name VARCHAR(255), user_id INT, created TIMESTAMP, description TEXT, deleted BOOLEAN, paused BOOLEAN, killed BOOLEAN, buildJob BOOLEAN, disk_size BIGINT, total_pairs INT, completed_pairs INT, errored_pairs INT, pending_pairs INT, status_code INT, max_stages INT, job_type INT, timeout INT, seed INT, suppress_output BOOLEAN, node_queued BOOLEAN) AS $$
+RETURNS TABLE(id INT, name VARCHAR(255), user_id INT, created TIMESTAMP, description TEXT, deleted BOOLEAN, paused BOOLEAN, killed BOOLEAN, buildJob BOOLEAN, disk_size BIGINT, total_pairs INT, completed_pairs INT, errored_pairs INT, pending_pairs INT, status_code INT, max_stages INT, job_type INT, timeout INT, seed BIGINT, suppress_output BOOLEAN, node_queued BOOLEAN) AS $$
 BEGIN
     RETURN QUERY
     SELECT
